Refactor: centralize game-specific theme providers
This commit introduces `SudokuGameTheme` to encapsulate CompositionLocals specific to the game feature, such as `LocalSudokuBoardColors`, `LocalKeyPadColors`, and `LocalHintLogsColors`.

**Key Changes:**

- **`SudokuGameTheme.kt` (feature/game):**
    - Created a new composable `SudokuGameTheme` that provides game-specific color schemes (`BoardColors`, `KeypadColors`, `HintLogsColors`) via CompositionLocal.
    - It retrieves the `LocalAppColorScheme` and uses it to remember these specific color sets.
    - It can optionally wrap its content with `SudokuSlayerTheme` or be used standalone if `SudokuSlayerTheme` is already provided by an ancestor.
- **`SudokuSlayerTheme.kt` (feature/uicore):**
    - Renamed the Material `colorScheme` variable to `materialColorScheme` to avoid ambiguity.
    - Now provides `LocalAppColorScheme` directly.
- **`SudokuGameScreen.kt` (feature/game):**
    - Now uses `SudokuGameTheme(useSudokuSlayerTheme = false)` to provide game-specific colors, assuming `SudokuSlayerTheme` is handled by the navigation host.
    - Removed direct instantiation and provision of `LocalSudokuBoardColors`, `LocalKeyPadColors`, and `LocalHintLogsColors`.
- **`AppContent.kt` (app):**
    - Removed the explicit `CompositionLocalProvider` for `LocalAppColorScheme` as it's now handled within `SudokuSlayerTheme`.

**Rationale:**

This change centralizes the provisioning of theme-related CompositionLocals.
- `SudokuSlayerTheme` in `uicore` provides the base application color scheme (`LocalAppColorScheme`) and Material theme.
- `SudokuGameTheme` in `feature/game` consumes the `LocalAppColorScheme` and derives and provides game-specific color sets.
This improves modularity and makes it clearer where different theme aspects are defined and provided.
